# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-oriented Go Web API scaffold. Application error codes: 0 success; 10000-19999 common/input; 20000-29999 authentication/authorization; 30000-39999 business; 50000-59999 infrastructure.",
        "title": "Go API Template",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/imports/cancel": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Cancel an import with optimistic locking",
                "parameters": [
                    {
                        "description": "Job and current version",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.VersionedImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/complete-upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Verify upload and enqueue validation",
                "parameters": [
                    {
                        "description": "Uploaded object metadata",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.CompleteUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/confirm": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Confirm a fully validated import",
                "parameters": [
                    {
                        "description": "Confirmation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ConfirmImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Create an import job and upload URL",
                "parameters": [
                    {
                        "description": "Import definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.CreateImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/datasets/describe": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Describe an available import dataset",
                "parameters": [
                    {
                        "description": "Provider and dataset",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.DescribeImportDatasetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/datasets/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Search available import datasets",
                "parameters": [
                    {
                        "description": "Search and pagination",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ListImportDatasetsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/error-report": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Create a validation error report URL",
                "parameters": [
                    {
                        "description": "Download request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrorReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/get": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Get an import job",
                "parameters": [
                    {
                        "description": "Job selector",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ImportSelector"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Search import jobs",
                "parameters": [
                    {
                        "description": "Filters and pagination",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ListImportsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/imports/retry": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "PSK": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "Upload a corrected file and retry validation",
                "parameters": [
                    {
                        "description": "Retry request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.RetryImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/version": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Return build and runtime version information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/buildinfo.Info"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/live": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Check process liveness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/health.Status"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ready": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Check database and Redis readiness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/health.Status"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httptransport.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/health.Status"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "buildinfo.Info": {
            "type": "object",
            "properties": {
                "build_time": {
                    "type": "string"
                },
                "commit": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "health.Dependency": {
            "type": "object",
            "properties": {
                "latency": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "health.Status": {
            "type": "object",
            "properties": {
                "dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/health.Dependency"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "httptransport.CompleteUploadRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "source_bytes": {
                    "type": "integer"
                },
                "source_checksum": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httptransport.ConfirmImportRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httptransport.CreateImportRequest": {
            "type": "object",
            "properties": {
                "dataset_code": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "provider_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.DescribeImportDatasetRequest": {
            "type": "object",
            "properties": {
                "dataset_code": {
                    "type": "string"
                },
                "provider_service": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ErrorReportRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "ttl_seconds": {
                    "type": "integer"
                }
            }
        },
        "httptransport.ImportSelector": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ListImportDatasetsRequest": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.ListImportsRequest": {
            "type": "object",
            "properties": {
                "created_from": {
                    "type": "string"
                },
                "created_to": {
                    "type": "string"
                },
                "dataset_code": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.Response": {
            "type": "object",
            "properties": {
                "body": {},
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "httptransport.RetryImportRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httptransport.VersionedImportRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter \"Bearer {JWT}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PSK": {
            "description": "Enter \"PSK {shared-key}\" for routes configured with PSK authentication.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}